requirements.txt
pysnmplib==5.0.21
snmpclitools==0.6.4
pyyaml==6.0
